+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Made all values in the `Std::PunchedArray` namespace private (since they are not intended to be used directly from outside).
 - You can now refer to entities using absolute namespace syntax (e.g., `::Std::String`) without importing them.
 - Changed the type of `Std::FFI::Destructor::make` function to return `IO`. Correspondingly, `Std::IO::IOHandle::from_file_ptr` also returns `IO`.
+- `Array::empty` and `Array::fill` now verifies the capacity and size arguments at runtime to ensure they are non-negative, and raises an error if they are negative.
 
 ### Fixed